requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
mistune==3.1.3
pytest==8.3.4
pytest-cov==6.0.0
